Major upgrade for getting attributes for SmartExpression.
If the attribute belongs to MathTex but not SmartExpression, it will run the lazy init_MathTex before accessing it so that everything will work properly. This also allows us to delete the .add monkeypatching.

If the attribute does not exist, it checks if that name is a SmartAction subclass, and if so, it applies that action's .get_output_expression to the expression and returns the resulting expression. Consequently these can be chained together.

This is honestly fucking amazing. Wow.

Next I will tackle .animate for action names in a similar way!